requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
dateparser==1.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
numpy==1.23.3
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
tzdata==2022.2
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
